Add duration timer to run page
